Renamed tests to "parameters"
