+++ Tests/EyeTrackerParametersTests.cpp
@@ -1,5 +1,5 @@
 #include "Data/EyeTracker.h"
-#include "XmlTestsShared.h"
+#include "ParametersTestsShared.h"
 
 using namespace qualisys_cpp_sdk::tests;
