+++ Tests/GeneralParametersTests.cpp
@@ -1,5 +1,5 @@
 #include "Data/General.h"
-#include "XmlTestsShared.h"
+#include "ParametersTestsShared.h"
 
 using namespace qualisys_cpp_sdk::tests;
