+++ Tests/ImageParametersTests.cpp
@@ -1,5 +1,5 @@
 #include "Data/Image.h"
-#include "XmlTestsShared.h"
+#include "ParametersTestsShared.h"
 
 using namespace qualisys_cpp_sdk::tests;
